// src/shared/domain/database-types.ts
export interface TagConfig {
  tagTable: string;
  relationTable: string;
  itemIdColumn: string;
  tagIdColumn: string;
  getTags: () => Promise<string[]>;
  groupTable: string;
}

/**
 * 标签配置映射
 */
export interface TagConfigMap {
  prompt: TagConfig;
  image: TagConfig;
}